mobilenet-model/requirements.txt
tensorflow>=2.4.0
numpy>=1.19.2
Pillow>=8.0.0
matplotlib>=3.3.0
argparse>=1.4.0